Some other changes, new class inverted index word search and word categorizer extracted in two classes.
