+++ phase02_clean_code/src/main/java/com/example/NecessaryFileNameCategorizer.java
@@ -6,10 +6,11 @@ import java.util.Set;
 
 public class NecessaryFileNameCategorizer {
     public Set<String> categorizer(InvertedIndex invertedIndex, ArrayList<String> words){
-        Set<String> finalResult = new HashSet<String>();
+        InvertedIndexWordSearch invertedIndexWordSearch = new InvertedIndexWordSearch(invertedIndex);
 
+        Set<String> finalResult = new HashSet<String>();
         for (String word : words) {
-            Set<String> searchResult = invertedIndex.searchWord(word);
+            Set<String> searchResult = invertedIndexWordSearch.searchWord(word);
             if(finalResult.isEmpty()){
                 finalResult.addAll(searchResult);
             }
