+++ phase02_clean_code/src/main/java/com/example/UnnecessaryFileNameCategorizer.java
@@ -8,10 +8,11 @@ import java.util.Set;
 public class UnnecessaryFileNameCategorizer {
 
     public Set<String> categorizer(InvertedIndex invertedIndex, ArrayList<String> words){
-        Set<String> finalResult = new HashSet<String>();
+        InvertedIndexWordSearch invertedIndexWordSearch = new InvertedIndexWordSearch(invertedIndex);
 
+        Set<String> finalResult = new HashSet<String>();
         for (String word : words) {
-            Set<String> searchResult = invertedIndex.searchWord(word);
+            Set<String> searchResult = invertedIndexWordSearch.searchWord(word);
                 finalResult.addAll(searchResult);
         }
